requirement.txt
torch==2.0.1
torchvision==0.15.2
onnxruntime==1.15.1
onnxruntime-gpu==1.15.1
opencv-python
fastapi
numpy
faiss-gpu
